Add Support for mediumtext, tinytext, geometry as text and import non UTC dbs
geometry is not supported by jhipster core, workarounds can be done to support it, most of these reolved around import as text and manually change back to Geometry or Point datatype.
added quartz db tables to the ignore list.
added serverTimezone=UTC to ensure reading database in other timezones does not throw error
tweaked pom.xml for database details to be in the properties section instead of split between properties and the generator